+++ src/components/common/button/OpenSheetBtn/OpenSheetBtn.stories.tsx
@@ -1,23 +1,19 @@
-import BottomSheet from '@/components/common/bottomSheet/BottomSheet';
-import OpenSheetBtnWithLabel from '@/components/housework/OpenSheetBtn/OpenSheetBtnWithLabel';
 import type { Meta, StoryObj } from '@storybook/react';
+import OpenSheetBtn from '@/components/common/button/OpenSheetBtn/OpenSheetBtn';
+import BottomSheet from '@/components/common/bottomSheet/BottomSheet';
 import { useState } from 'react';
 
 const meta = {
-  title: 'components/housework/OpenSheetBtn/OpenSheetBtnWithLabel',
-  component: OpenSheetBtnWithLabel,
+  title: 'components/common/button/OpenSheetBtn/OpenSheetBtn',
+  component: OpenSheetBtn,
   tags: ['autodocs'],
   argTypes: {
-    title: {
-      description: '타이틀',
-      control: 'text',
-    },
-    selected: {
-      description: '셀렉된 값',
+    text: {
+      description: '셀렉트버튼 제목',
       control: 'text',
     },
   },
-} satisfies Meta<typeof OpenSheetBtnWithLabel>;
+} satisfies Meta<typeof OpenSheetBtn>;
 
 export default meta;
 
@@ -26,11 +22,7 @@ const SelectBtnWithBottomSheet = () => {
 
   return (
     <>
-      <OpenSheetBtnWithLabel
-        title='타이틀'
-        selected='셀렉된 값'
-        handleClick={() => setOpen(true)}
-      />
+      <OpenSheetBtn text='셀렉트버튼 제목' handleClick={() => setOpen(true)} type='' />
       <BottomSheet isOpen={isOpen} setOpen={setOpen} title='바텀시트 제목'>
         <div className='p-4'>
           <p>바텀시트 컨텐츠입니다</p>
