+++ src/pages/HouseWorkStepTwoPage.tsx
@@ -1,8 +1,8 @@
 import Button from '@/components/common/button/Button/Button';
 import HeaderWithTitle from '@/components/housework/HeaderWithTitle/HeaderWithTitle';
 import ManagerSelectSheet from '@/components/housework/ManagerSelectSheet/ManagerSelectSheet';
-import OpenSheetBtn from '@/components/housework/OpenSheetBtn/OpenSheetBtn';
-import OpenSheetBtnWithLabel from '@/components/housework/OpenSheetBtn/OpenSheetBtnWithLabel';
+import OpenSheetBtn from '@/components/common/button/OpenSheetBtn/OpenSheetBtn';
+import OpenSheetBtnWithLabel from '@/components/common/button/OpenSheetBtn/OpenSheetBtnWithLabel';
 import TimeControl from '@/components/housework/TimeControl/TimeControl';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
